+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added:
+- **SlashCreator:** Add client passthrough option
 ## [4.4.2] - 2021-12-20
 ### Changed:
 - Minimum Node Version is now v14